fix(ci): split two files back under the 500-LoC cap I pushed them over
`Enforce ≤500 LoC per source file` is a REQUIRED check and I broke it. Both
files were under the cap on main and my own commits pushed them over:

    benchmarks/agentic/mod.rs      493 -> 567   (7 lines of headroom, gone)
    gate/coverage_map_tests.rs     468 -> 573

Split along the seams the code already had, not arbitrarily:

* `score::per_step_tallies` and `score::verdict` move to `score.rs`. Both are
  SCORING; `mod.rs` drives iterations. That is the same separation
  `serve_matrix` uses, and the verdict is now readable without paging through
  sandbox setup.
* The canonical serve command and the `--mtp-gate` rationale move to
  `descriptors.rs`, beside the thresholds they justify. They were duplicated
  in two places, which is how they drift.
* The promotion-debt and boundary tests move to `coverage_promotion_tests.rs`.
  Different question from `coverage_map_tests`: that file asks what the
  REQUIRED gates invalidate, this asks what a gate that is not required yet
  would still have wanted to see.

★ I got the rule wrong twice on the way here, both times by assuming instead
of reading. First I "verified" the fix by checking whether files anywhere in
the tree exceeded 500 — and found `build.rs` at 1270 and `traits/model.rs` at
1180, which nearly convinced me the check could not mean what it says. It
applies to `crates/**/*.rs` with a documented allow-list, and those entries
are ON it. Second, an over-eager extraction took the impl block's closing brace
with it and produced a file that did not parse; reverted and redone with
explicit anchors.

454 tests green; fmt clean; clippy 0; every touched file under the cap.
