+++ crates/atlas-plugin/src/benchmarks/agentic/mod.rs
@@ -23,32 +23,10 @@
 //! comparable to that band when served exactly like this (no docker wrapper —
 //! run the binary directly):
 //!
-//! ```sh
-//! spark serve Qwen/Qwen3.6-35B-A3B-FP8 \
-//!   --max-seq-len 32768 --max-batch-size 2 --gpu-memory-utilization 0.70 \
-//!   --kv-cache-dtype bf16 --lm-head-dtype bf16 --scheduling-policy slai \
-//!   --speculative --num-drafts 1 --mtp-quantization bf16 --mtp-gate force \
-//!   --enable-prefix-caching --ssm-cache-slots 256 --ssm-checkpoint-interval 16 \
-//!   --kv-high-precision-layers auto --port 8888
-//! ```
+//! The canonical serve command, and why `--mtp-gate force` belongs in it,
+//! are documented on `DESCRIPTOR` in `descriptors.rs` — beside the thresholds
+//! they justify, rather than duplicated here where they would drift.
 //!
-//! ★★ THIS BLOCK IS DOCUMENTATION. Nothing executes it.
-//!
-//! Under `--pull-request-gate` the serve is built by `bench_selfstart` from the
-//! RECIPE named in `BENCH.toml` — `qwen3.6/qwen3.6-35b-a3b-fp8-bf16head`, which
-//! lives in the separate `atlas-recipes` repo and is honoured verbatim. Editing
-//! the command below changes what a reader believes, not what runs. It is
-//! written down here because it is the shape an operator reproduces by hand,
-//! and it must not drift from the recipe.
-//!
-//! ★ `--mtp-gate force` is a DETERMINISM pin, and its absence is the root cause
-//! of this gate's intermittent 9/10 on `followed_directions`.
-//!
-//! **IT IS NOT YET IN EFFECT.** The recipe carries `speculative: true` and
-//! `mtp_quantization: bf16` and no `mtp_gate` key; `--mtp-gate` is
-//! `Option<String>` with no clap default, so absent means `auto`. Closing this
-//! requires a PR to `atlas-recipes` adding `mtp_gate: force` to that recipe.
-//! Until that lands, the flip described below can still happen.
 //!
 //! In `auto`, the MTP gate is a bandit arbiter that switches MTP<->serial at
 //! runtime on **wall-clock** tok/s EWMAs. Speculation is NOT output-neutral at
@@ -71,14 +49,7 @@
 //! the flip — was explicitly rejected: it would convert a determinism bug into
 //! permanent noise the gate can never see through again.
 //!
-//! Two further pins that must not drift:
-//!
-//! * **`--lm-head-dtype bf16` is mandatory.** fp8 and nvfp4 lm-heads pass
-//!   short-prompt coherence but degenerate over this harness's long structured
-//!   trajectories (low-margin argmax flips compound turn over turn).
-//! * **Thinking stays ON** — no `--disable-thinking`. The historical band was
-//!   recorded with thinking enabled; disabling it is the TTFT/BFCL-leg recipe
-//!   (their probes count only content tokens), not Gate A's.
+//! Two further pins that must not drift (see `descriptors.rs`).
 
 pub mod agent;
 pub mod preflight;
@@ -306,54 +277,14 @@ impl AgenticWebserver {
         );
         m.insert("sum_wall_s".to_string(), self.total_wall());
 
-        // ★ Per-directive tallies, so the RECORD can answer "which one failed"
-        // without the trajectory — which the next run of the same index
-        // truncates. `followed_directions` above is all-or-nothing per
-        // iteration; these say how many iterations evidenced each step.
-        //
-        // Gate bounds are opt-in per metric (`BENCH.toml`), so adding keys
-        // cannot tighten or loosen anything on its own: an unbounded metric is
-        // recorded and reported, never compared. That is what makes it safe to
-        // record more than is gated.
-        for (name, _) in self
-            .rows
-            .first()
-            .map(|r| r.directions.steps.clone())
-            .unwrap_or_default()
-        {
-            let met = self
-                .rows
-                .iter()
-                .filter(|r| r.directions.steps.iter().any(|(n, ok)| *n == name && *ok))
-                .count();
-            m.insert(format!("step:{name}"), met as f64);
-        }
+        m.extend(score::per_step_tallies(
+            &self.rows.iter().map(|r| &r.directions).collect::<Vec<_>>(),
+        ));
         m
     }
 
     fn verdict(&self) -> Verdict {
-        let n = self.rows.len();
-        let ok = self.rows.iter().filter(|r| r.webserver_ok).count();
-        let fd = self.rows.iter().filter(|r| r.directions.overall()).count();
-        let wall = self.total_wall();
-        let mut failures = Vec::new();
-        if ok < n {
-            failures.push(format!("webserver_ok {ok}/{n}"));
-        }
-        if fd < n {
-            failures.push(format!("followed_directions {fd}/{n}"));
-        }
-        if wall > self.wall_budget_s {
-            failures.push(format!("Σwall {wall:.0}s > {:.0}s", self.wall_budget_s));
-        }
-        if failures.is_empty() {
-            Verdict::pass(format!(
-                "{ok}/{n} webserver_ok · {fd}/{n} followed_directions · Σwall {wall:.0}s ≤ {:.0}s",
-                self.wall_budget_s
-            ))
-        } else {
-            Verdict::fail(failures.join(" · "))
-        }
+        score::verdict(&self.rows, self.total_wall(), self.wall_budget_s)
     }
 }
